{"schema":"decodex/commit/1","summary":"Move keyboard input runtime","authority":"manual"}
